Fix Content-Length for UTF-8 messages
String.length returns invalid count for UTF-8 strings. This causes the POST message to be truncated because the receiving server stops at Content-length's count. WNS will reply with "received" but the message is discarded as it will be containing incomplete XML code.

Using the Buffer global class, the correct Content-length can be calculated.